Исправил отображение кнопок в замерах и фото. Добавил колонки с заменой текста в фотографиях. Добавил фильтр по марафону в пользовательские данные. Изменил структуру проекта. Исправил недочеты и ошибки.

+++ marathon_bot/scenarios/handler_register.py
@@ -12,43 +12,51 @@ from marathon_bot.states.all_states_menu import MainMenu
 from marathon_bot.states.state_scenarios import Register
 
 
-@db_session
 async def send_welcome(message: types.Message, action='send'):
-    try:
-        if not AllUsers.get(tg_id=message.from_user.id):
-            AllUsers(tg_id=message.from_user.id)
-            commit()
-    except MultipleObjectsFoundError:
-        pass
-    marathons = Marathon.select().order_by(Marathon.id)[:]
-    markup = types.InlineKeyboardMarkup()
-    users = Users.select().where(tg_id=message.chat.id)[:]
-    for marathon in marathons:
-        if not marathon.date_start <= datetime.datetime.now().date() < marathon.date_end:
-            continue
-        text = f"{marathon.name}"
-        if any([user for user in users if user.marathon == marathon]):
-            text += " ✅️"
-        if marathon.price > 0:
-            text += " 💎"
-        else:
-            text += " 🆓"
-        if marathon.close:
-            text += " 🛠"
-        markup.add(
-            types.InlineKeyboardButton(
-                text=text,
-                callback_data=marathon.name + '_marathon')
-        )
+    with db_session:
+        try:
+            if not AllUsers.get(tg_id=message.from_user.id):
+                AllUsers(tg_id=message.from_user.id)
+                commit()
+        except MultipleObjectsFoundError:
+            pass
+        marathons = Marathon.select().order_by(Marathon.id)[:]
+        markup = types.InlineKeyboardMarkup()
+        users = Users.select().where(tg_id=message.chat.id)[:]
+        for marathon in marathons:
+            if marathon.date_start <= datetime.datetime.now().date() <= marathon.date_end:
+                text = f"{marathon.name}"
+                if marathon.close:
+                    text += " 🛠"
+                if any([user for user in users if user.marathon == marathon]):
+                    text += " ✅️"
+                if marathon.price > 0:
+                    if marathon.count_users <= 0:
+                        if ' ✅️' in text:
+                            pass
+                        else:
+                            continue
+                    text += " 💎"
+                    text += f" Осталось {marathon.count_users} мест!"
+                else:
+                    text += " 🆓"
+                markup.add(
+                    types.InlineKeyboardButton(
+                        text=text,
+                        callback_data=marathon.name + '_marathon')
+                )
+    if message.content_type == 'invoice' and action == 'edit':
+        await message.delete()
+        action = 'send'
     if action == 'send':
         await message.answer("Добро пожаловать!\nСписок доступных на данный момент марафонов:\n\n"
                              "🛠 - марафон на техническом обслуживании\n"
-                             "💎/🆓 - платный/бесплатный марафон\n"
+                             # "💎/🆓 - платный/бесплатный марафон\n"
                              "✅ - вы зарегистрированы в марафоне\n", reply_markup=markup)
     elif action == 'edit':
         await message.edit_text("Добро пожаловать!\nСписок доступных на данный момент марафонов:\n\n"
                                 "🛠 - марафон на техническом обслуживании\n"
-                                "💎/🆓 - платный/бесплатный марафон\n"
+                                # "💎/🆓 - платный/бесплатный марафон\n"
                                 "✅ - вы зарегистрированы в марафоне\n", reply_markup=markup)
     await Register.choice_marathon.set()
 
@@ -58,14 +66,14 @@ async def check_register_from_marathon(query: types.CallbackQuery, state: FSMCon
     cur = con.cursor(cursor_factory=NamedTupleCursor)
     cur.execute(
         f"""
-            SELECT pam.id
+            SELECT pam.id, pau.first_name, pau.last_name
             FROM marathon as pam
             INNER JOIN users as pau ON pam.id = pau.marathon_id
             WHERE pau.tg_id = {query.from_user.id} AND pam.name = \'{query.data.split("_marathon")[0]}\'
         """
     )
     marathon = cur.fetchone()
-    if not marathon:
+    if not marathon or (marathon.first_name == '.' and marathon.last_name == '.'):
         markup = types.InlineKeyboardMarkup()
         markup.add(
             types.InlineKeyboardButton(
@@ -99,9 +107,22 @@ async def none_register_marathon(query: types.CallbackQuery):
 async def register_marathon(query: types.CallbackQuery, state: FSMContext):
     data = await state.get_data()
     marathon = await Marathon.get_marathon(marathon_id=data['marathon_id'])
-    if marathon.price > 0.0:
+    if marathon.count_users <= 0:
+        return await none_register_marathon(query)
+    try:
+        user = await Users.get_user(query.from_user.id, data['marathon_id'])
+        check = True if user.is_pay else False
+    except Exception:
+        check = False
+    if marathon.price > 0.0 and not check:
         prices = [types.LabeledPrice(label=marathon.name, amount=marathon.price * 100)]
         await query.message.delete()
+        markup = types.InlineKeyboardMarkup()
+        markup.add(types.InlineKeyboardButton(text=f'Оплатить {marathon.price} РУБ', pay=True)).add(
+            types.InlineKeyboardButton(
+                text='Назад',
+                callback_data='None')
+        )
         await bot.send_invoice(
             query.message.chat.id,
             title=marathon.name,
@@ -113,9 +134,21 @@ async def register_marathon(query: types.CallbackQuery, state: FSMContext):
             need_email=True,
             need_phone_number=True,
             start_parameter='time-machine-example',
-            payload='register-marathon-payments'
+            payload='register-marathon-payments',
+            reply_markup=markup
         )
     else:
+        state_data = await state.get_data()
+        Users(
+            tg_id=query.from_user.id,
+            username=query.from_user.username,
+            first_name='.',
+            last_name='.',
+            scopes=0,
+            marathon=state_data['marathon_id'],
+            is_pay=False,
+        )
+        commit()
         msg = await query.message.edit_text(
             "Доброго времени суток!\nПрошу заполнять данные верно, так как в дальнейшем они будут использоваться "
             "для подсчета данных!\n Как вас зовут?(Напишите ваши Фамилию и Имя через пробел)\n"
@@ -130,6 +163,17 @@ async def process_pre_checkout_query(pre_checkout_query: types.PreCheckoutQuery)
 
 
 async def process_successful_payment(message: types.Message, state: FSMContext):
+    state_data = await state.get_data()
+    with db_session:
+        Users(
+            tg_id=message.chat.id,
+            username=message.chat.username,
+            first_name='.',
+            last_name='.',
+            scopes=0,
+            marathon=state_data['marathon_id'],
+            is_pay=True,
+        )
     msg = await message.answer(
         "Доброго времени суток!\nПрошу заполнять данные верно, так как в дальнейшем они будут использоваться "
         "для подсчета данных!\n Как вас зовут?(Напишите ваши Фамилию и Имя через пробел)\n"
@@ -142,12 +186,11 @@ async def process_successful_payment(message: types.Message, state: FSMContext):
 async def get_full_name(message: types.Message, state: FSMContext):
     state_data = await state.get_data()
     try:
-        sql = "insert into users(tg_id, username, first_name, last_name, scopes, marathon_id) values " \
-              f"(\'{message.from_user.id}\', \'{message.from_user.username if message.from_user.username else '-'}\'," \
-              f"\'{message.text.split(' ')[1]}\', \'{message.text.split(' ')[0]}\', 0, {state_data['marathon_id']})"
-        cur = con.cursor()
-        cur.execute(sql)
-        con.commit()
+        with db_session:
+            user = await Users.get_user(message.chat.id, state_data['marathon_id'])
+            user.first_name = message.text.split(' ')[1]
+            user.last_name = message.text.split(' ')[0]
+            user.marathon.count_users -= 1
     except Exception as exc:
         failure_text = "Что-то пошло не так. Повторите попытку! (Напишите ваши Фамилию и Имя через пробел)\n" \
                        "❗️Например: Иванова Мария"
